requirement.txt
pytest
wheel
pyyaml
allure-pytest